+++ R/pseudo-id.R
@@ -14,20 +14,22 @@
 #' attr(pid, "uniques")[pid]
 #'
 #' @export
-pseudo_id <- function(x) {
+pseudo_id <- function(x, ...) {
   UseMethod("pseudo_id", x)
 }
 
 #' @export
 #' @rdname pseudo_id
-pseudo_id.pseudo_id <- function(x) {
+pseudo_id.pseudo_id <- function(x, ...) {
   x
 }
 
 #' @export
 #' @rdname pseudo_id
-pseudo_id.default <- function(x) {
-  ux <- na_last(unique(x))
+#' @param na_last `Logical` if `FALSE` will not place `NA` at the end
+pseudo_id.default <- function(x, na_last = TRUE, ...) {
+  ux <- unique(x)
+  if (na_last) ux <- na_last(ux)
   make_pseudo_id(match(x, ux), ux)
 }
 
@@ -42,7 +44,7 @@ pseudo_id.factor <- function(x) {
 #' @export
 print.pseudo_id <- function(x, ...) {
   print(as.integer(x))
-  cat("Uniques: ", paste0(unique(x), sep = " "), "\n", sep = "")
+  cat("Uniques: ", paste0(attr(x, "uniques"), sep = " "), "\n", sep = "")
   invisible(x)
 }