+++ R/unlist.R
@@ -33,5 +33,8 @@ unlist0 <- function(x) {
 #' @param sep A separation for combining names
 squash_vec <- function(x, sep = ".") {
   x <- unlist0(x)
-  unique(x) %names% mark::vap_chr(split(names(x), x), collapse0, sep = sep)
+  id <- pseudo_id(x, na_last = FALSE)
+  nm <- names(x)
+  squasher <- function(i) collapse0(nm[i], sep = sep)
+  .uniques(id) %names% vap_chr(split(seq_along(id), id), squasher)
 }