+++ tests/testthat/scripts/lines/lines.R
@@ -1,0 +1,10 @@
+
+"Hello, world"
+
+# how many lines are in here?
+
+
+# the number of lines is ...
+
+
+# 10?