+++ tests/testthat/test-lines-of-code.R
@@ -1,0 +1,8 @@
+test_that("lines_of_r_code() works", {
+  x <- test_path("scripts/lines")
+  expect_equal(lines_of_r_code(x), 4L)
+  expect_equal(lines_of_r_code(x, skip_empty = FALSE), 10L)
+
+  x <-  test_path("scripts/lines/lines.R")
+  expect_equal(lines_of_r_code(x), 4L)
+})