+++ NEWS.md
@@ -1,3 +1,5 @@
+# mark (development version)
+
 # mark 0.3.0
 
 ## Fixes