+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * adds `percentile_rank()` to calculate percentile ranks with a vector
 * `pseudo_id()` gains argument `na_last` to change positioning of `NA` values
+* adds `insert()` to insert multiple values into a vector
 
 # mark 0.3.0