+++ NEWS.md
@@ -5,6 +5,7 @@
 * adds `percentile_rank()` to calculate percentile ranks with a vector
 * adds `insert()` to insert multiple values into a vector
 * `pseudo_id()` gains argument `na_last` to change positioning of `NA` values
+* `is_true()` and `is_false()` are now exported as generics with methods for `default` and `logical`
 
 ## Fixes