+++ NEWS.md
@@ -1,4 +1,4 @@
-# mark (development version)
+# mark 0.4.0
 
 ## New features